Adicionando classe de teste

+++ src/main/kotlin/one/digitalinnovation/digionebank/Conta.kt
@@ -1,17 +1,18 @@
 package one.digitalinnovation.digionebank
 
+import java.math.BigDecimal
 
 
 class Conta(
     val agencia: String,
     val numero: String,
-    val saldo: Double
+    val saldo: BigDecimal
 ) {
-    fun deposito(valor:Double){
+    fun deposito(valor:BigDecimal){
 
     }
 
-    fun saque(valor:Double){
+    fun saque(valor:BigDecimal){
 
     }
